Move mock code to own project
