+++ Echo.sln
@@ -58,6 +58,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Echo.Platforms.Dnlib.Tests"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Echo.Ast.Tests", "test\Core\Echo.Ast.Tests\Echo.Ast.Tests.csproj", "{C514B504-1694-4F5E-9721-DB85C15AE474}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mocks", "test\Platforms\Mocks\Mocks.csproj", "{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -148,6 +150,10 @@ Global
 		{C514B504-1694-4F5E-9721-DB85C15AE474}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C514B504-1694-4F5E-9721-DB85C15AE474}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C514B504-1694-4F5E-9721-DB85C15AE474}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7C8344E1-C809-412F-9284-13F98345DF27} = {8344861B-4B03-495F-A6B5-54BF6EF37B4C}
@@ -176,5 +182,6 @@ Global
 		{CE527A7F-3D42-48B3-A9DB-24ED24C8F134} = {895F6ABB-3011-4F65-A814-046B5C1B2DD0}
 		{9AF949E7-3C17-4DC3-ABB3-A25BE986EB31} = {8344861B-4B03-495F-A6B5-54BF6EF37B4C}
 		{C514B504-1694-4F5E-9721-DB85C15AE474} = {91D56580-50E2-49B4-A5C0-DB83C9AADF8D}
+		{8CF58E4A-D381-4FFD-B6CA-85C346BD9AA8} = {895F6ABB-3011-4F65-A814-046B5C1B2DD0}
 	EndGlobalSection
 EndGlobal
