+++ test/Platforms/Echo.Platforms.AsmResolver.Tests/Emulation/Dispatch/Pointers/LdObjTest.cs
@@ -6,6 +6,7 @@ using Echo.Concrete.Values.ValueType;
 using Echo.Platforms.AsmResolver.Emulation;
 using Echo.Platforms.AsmResolver.Emulation.Values.Cli;
 using Echo.Platforms.AsmResolver.Tests.Mock;
+using Mocks;
 using Xunit;
 
 namespace Echo.Platforms.AsmResolver.Tests.Emulation.Dispatch.Pointers
