+++ test/Platforms/Echo.Platforms.Dnlib.Tests/Echo.Platforms.Dnlib.Tests.csproj
@@ -15,6 +15,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\..\..\src\Platforms\Echo.Platforms.Dnlib\Echo.Platforms.Dnlib.csproj" />
+        <ProjectReference Include="..\Mocks\Mocks.csproj" />
     </ItemGroup>
 
 </Project>
