+++ test/Platforms/Echo.Platforms.Dnlib.Tests/StaticSuccessorResolverTest.cs
@@ -1,7 +1,6 @@
 using System.Linq;
 using dnlib.DotNet;
-using dnlib.DotNet.Emit;
-using Echo.Platforms.Dnlib.Tests.Mock;
+using Mocks;
 using Xunit;
 
 namespace Echo.Platforms.Dnlib.Tests
